add response advice
